Create HTML for first section of webpage
The HTML content has been created for the first section of the
webpage. The "first section" refers to the dark gray section
seen in the reference photo.

As stated in the README file, this version of the site will
not have the exact same content as the reference photo. There will
be a general theme of "A Song of Ice and Fire".